Generate Confidence Intervals for the Total, Direct, and Indirect Effects

The function generates confidence intervals for the total, direct, amd indirect effects using the cTMed package.

Usage

CI(x, delta_t = 1:24, R = 20000L)

Arguments

x

R object. Output of the FitDynr() function.

delta_t

Numeric vector. Vector of time-intervals to use.

R

Positive integer. Number of Monte Carlo replications.

See also

Examples

# dynr
library(dynr)
#> Loading required package: ggplot2
data <- Data(repid = 1, n = 50)
fit <- FitDynr(x = data)
#> [1] "Get ready!!!!"
#> using C compiler: ‘gcc (Ubuntu 11.4.0-1ubuntu1~22.04) 11.4.0’
#> Optimization function called.
#> Starting Hessian calculation ...
#> Finished Hessian calculation.
#> Original exit flag:  3 
#> Modified exit flag:  3 
#> Optimization terminated successfully: ftol_rel or ftol_abs was reached. 
#> Original fitted parameters:  -0.2605694 -0.0743651 0.08194417 0.6720199 
#> -0.460244 -0.02385756 -0.4523976 0.7365811 -0.6910508 -1.475439 0.2015311 
#> -0.2468988 -2.729321 0.2947053 -2.877131 -1.612013 -1.617904 -1.576812 
#> -0.1284087 0.06218073 0.2874795 -0.175533 0.2349609 -0.1969349 
#> 
#> Transformed fitted parameters:  -0.2605694 -0.0743651 0.08194417 0.6720199 
#> -0.460244 -0.02385756 -0.4523976 0.7365811 -0.6910508 0.2286783 0.04608579 
#> -0.0564604 0.07455129 0.007854991 0.07590425 0.1994856 0.198314 0.2066329 
#> 0.8794939 0.05468757 0.2528364 0.8424102 0.212856 0.9402482 
#> 
#> Doing end processing
#> Successful trial
#> Total Time: 3.814901 
#> Backend Time: 3.806875 
CI(fit)
#> $type
#> [1] "dynr"
#> 
#> $delta_t
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24
#> 
#> $delta
#> $`1`
#>          interval     est     se       z p    2.5%   97.5%
#> total           1 -0.1248 0.0282 -4.4232 0 -0.1801 -0.0695
#> direct          1 -0.2815 0.0376 -7.4927 0 -0.3552 -0.2079
#> indirect        1  0.1567 0.0168  9.3081 0  0.1237  0.1897
#> 
#> $`2`
#>          interval     est     se       z      p    2.5%   97.5%
#> total           2  0.0408 0.0310  1.3179 0.1875 -0.0199  0.1016
#> direct          2 -0.3515 0.0544 -6.4577 0.0000 -0.4582 -0.2448
#> indirect        2  0.3924 0.0452  8.6767 0.0000  0.3037  0.4810
#> 
#> $`3`
#>          interval     est     se       z p    2.5%   97.5%
#> total           3  0.2180 0.0313  6.9709 0  0.1567  0.2793
#> direct          3 -0.3302 0.0620 -5.3210 0 -0.4518 -0.2086
#> indirect        3  0.5481 0.0700  7.8349 0  0.4110  0.6853
#> 
#> $`4`
#>          interval     est     se       z p    2.5%   97.5%
#> total           4  0.3257 0.0348  9.3684 0  0.2576  0.3939
#> direct          4 -0.2765 0.0635 -4.3510 0 -0.4010 -0.1519
#> indirect        4  0.6022 0.0855  7.0449 0  0.4347  0.7697
#> 
#> $`5`
#>          interval     est     se       z     p    2.5%   97.5%
#> total           5  0.3629 0.0390  9.3003 0e+00  0.2865  0.4394
#> direct          5 -0.2177 0.0607 -3.5847 3e-04 -0.3368 -0.0987
#> indirect        5  0.5807 0.0912  6.3681 0e+00  0.4020  0.7594
#> 
#> $`6`
#>          interval     est     se       z      p    2.5%   97.5%
#> total           6  0.3519 0.0419  8.4027 0.0000  0.2698  0.4340
#> direct          6 -0.1651 0.0552 -2.9888 0.0028 -0.2733 -0.0568
#> indirect        6  0.5170 0.0891  5.8028 0.0000  0.3424  0.6916
#> 
#> $`7`
#>          interval     est     se       z      p    2.5%   97.5%
#> total           7  0.3153 0.0430  7.3243 0.0000  0.2309  0.3997
#> direct          7 -0.1220 0.0484 -2.5232 0.0116 -0.2168 -0.0272
#> indirect        7  0.4373 0.0821  5.3300 0.0000  0.2765  0.5981
#> 
#> $`8`
#>          interval     est     se       z      p    2.5%   97.5%
#> total           8  0.2693 0.0430  6.2580 0.0000  0.1849  0.3536
#> direct          8 -0.0886 0.0411 -2.1554 0.0311 -0.1692 -0.0080
#> indirect        8  0.3579 0.0727  4.9225 0.0000  0.2154  0.5004
#> 
#> $`9`
#>          interval     est     se       z      p    2.5%  97.5%
#> total           9  0.2234 0.0424  5.2713 0.0000  0.1403 0.3065
#> direct          9 -0.0635 0.0341 -1.8612 0.0627 -0.1304 0.0034
#> indirect        9  0.2869 0.0631  4.5467 0.0000  0.1632 0.4106
#> 
#> $`10`
#>          interval     est     se       z      p    2.5%  97.5%
#> total          10  0.1822 0.0414  4.4022 0.0000  0.1011 0.2633
#> direct         10 -0.0451 0.0278 -1.6231 0.1046 -0.0996 0.0094
#> indirect       10  0.2273 0.0545  4.1701 0.0000  0.1205 0.3341
#> 
#> $`11`
#>          interval     est     se       z      p    2.5%  97.5%
#> total          11  0.1473 0.0401  3.6715 0.0002  0.0687 0.2259
#> direct         11 -0.0318 0.0222 -1.4283 0.1532 -0.0754 0.0118
#> indirect       11  0.1790 0.0474  3.7741 0.0002  0.0861 0.2720
#> 
#> $`12`
#>          interval     est     se       z      p    2.5%  97.5%
#> total          12  0.1185 0.0385  3.0804 0.0021  0.0431 0.1940
#> direct         12 -0.0223 0.0176 -1.2671 0.2051 -0.0567 0.0122
#> indirect       12  0.1408 0.0419  3.3632 0.0008  0.0587 0.2229
#> 
#> $`13`
#>          interval     est     se       z      p    2.5%  97.5%
#> total          13  0.0953 0.0365  2.6137 0.0090  0.0238 0.1668
#> direct         13 -0.0155 0.0137 -1.1327 0.2574 -0.0424 0.0113
#> indirect       13  0.1109 0.0374  2.9606 0.0031  0.0375 0.1842
#> 
#> $`14`
#>          interval     est     se       z      p    2.5%  97.5%
#> total          14  0.0767 0.0341  2.2490 0.0245  0.0099 0.1435
#> direct         14 -0.0108 0.0106 -1.0194 0.3080 -0.0315 0.0100
#> indirect       14  0.0875 0.0338  2.5912 0.0096  0.0213 0.1537
#> 
#> $`15`
#>          interval     est     se       z      p    2.5%  97.5%
#> total          15  0.0618 0.0315  1.9637 0.0496  0.0001 0.1235
#> direct         15 -0.0075 0.0081 -0.9233 0.3558 -0.0233 0.0084
#> indirect       15  0.0693 0.0305  2.2693 0.0232  0.0094 0.1291
#> 
#> $`16`
#>          interval     est     se       z      p    2.5%  97.5%
#> total          16  0.0498 0.0287  1.7385 0.0821 -0.0063 0.1060
#> direct         16 -0.0052 0.0061 -0.8411 0.4003 -0.0172 0.0069
#> indirect       16  0.0550 0.0275  1.9979 0.0457  0.0010 0.1090
#> 
#> $`17`
#>          interval     est     se       z      p    2.5%  97.5%
#> total          17  0.0402 0.0258  1.5583 0.1192 -0.0104 0.0909
#> direct         17 -0.0036 0.0046 -0.7704 0.4411 -0.0126 0.0055
#> indirect       17  0.0438 0.0247  1.7726 0.0763 -0.0046 0.0922
#> 
#> $`18`
#>          interval     est     se       z      p    2.5%  97.5%
#> total          18  0.0325 0.0230  1.4117 0.1580 -0.0126 0.0777
#> direct         18 -0.0025 0.0035 -0.7090 0.4783 -0.0092 0.0043
#> indirect       18  0.0350 0.0220  1.5863 0.1127 -0.0082 0.0782
#> 
#> $`19`
#>          interval     est     se       z      p    2.5%  97.5%
#> total          19  0.0263 0.0204  1.2904 0.1969 -0.0136 0.0662
#> direct         19 -0.0017 0.0026 -0.6555 0.5121 -0.0067 0.0034
#> indirect       19  0.0280 0.0195  1.4317 0.1522 -0.0103 0.0663
#> 
#> $`20`
#>          interval     est     se       z      p    2.5%  97.5%
#> total          20  0.0213 0.0179  1.1885 0.2347 -0.0138 0.0564
#> direct         20 -0.0012 0.0019 -0.6086 0.5428 -0.0049 0.0026
#> indirect       20  0.0224 0.0172  1.3025 0.1928 -0.0113 0.0562
#> 
#> $`21`
#>          interval     est     se       z      p    2.5%  97.5%
#> total          21  0.0172 0.0156  1.1015 0.2707 -0.0134 0.0478
#> direct         21 -0.0008 0.0014 -0.5672 0.5706 -0.0036 0.0020
#> indirect       21  0.0180 0.0151  1.1934 0.2327 -0.0116 0.0476
#> 
#> $`22`
#>          interval     est     se       z      p    2.5%  97.5%
#> total          22  0.0139 0.0136  1.0265 0.3047 -0.0127 0.0405
#> direct         22 -0.0005 0.0010 -0.5305 0.5957 -0.0026 0.0015
#> indirect       22  0.0145 0.0131  1.1005 0.2711 -0.0113 0.0402
#> 
#> $`23`
#>          interval     est     se       z      p    2.5%  97.5%
#> total          23  0.0113 0.0117  0.9609 0.3366 -0.0117 0.0342
#> direct         23 -0.0004 0.0008 -0.4978 0.6186 -0.0019 0.0011
#> indirect       23  0.0116 0.0114  1.0206 0.3074 -0.0107 0.0340
#> 
#> $`24`
#>          interval     est     se       z      p    2.5%  97.5%
#> total          24  0.0091 0.0101  0.9031 0.3665 -0.0107 0.0289
#> direct         24 -0.0003 0.0005 -0.4686 0.6393 -0.0013 0.0008
#> indirect       24  0.0094 0.0098  0.9513 0.3414 -0.0099 0.0287
#> 
#> 
#> $mc
#> $`1`
#>          interval     est     se     R    2.5%   97.5%
#> total           1 -0.1248 0.0282 20000 -0.1806 -0.0706
#> direct          1 -0.2815 0.0375 20000 -0.3571 -0.2098
#> indirect        1  0.1567 0.0168 20000  0.1255  0.1918
#> 
#> $`2`
#>          interval     est     se     R    2.5%   97.5%
#> total           2  0.0408 0.0312 20000 -0.0224  0.0993
#> direct          2 -0.3515 0.0550 20000 -0.4692 -0.2526
#> indirect        2  0.3924 0.0455 20000  0.3108  0.4891
#> 
#> $`3`
#>          interval     est     se     R    2.5%   97.5%
#> total           3  0.2180 0.0314 20000  0.1567  0.2801
#> direct          3 -0.3302 0.0638 20000 -0.4745 -0.2250
#> indirect        3  0.5481 0.0712 20000  0.4272  0.7059
#> 
#> $`4`
#>          interval     est     se     R    2.5%   97.5%
#> total           4  0.3257 0.0348 20000  0.2612  0.3982
#> direct          4 -0.2765 0.0667 20000 -0.4348 -0.1755
#> indirect        4  0.6022 0.0882 20000  0.4594  0.8027
#> 
#> $`5`
#>          interval     est     se     R    2.5%   97.5%
#> total           5  0.3629 0.0393 20000  0.2935  0.4472
#> direct          5 -0.2177 0.0655 20000 -0.3797 -0.1263
#> indirect        5  0.5807 0.0957 20000  0.4337  0.8067
#> 
#> $`6`
#>          interval     est     se     R    2.5%   97.5%
#> total           6  0.3519 0.0424 20000  0.2796  0.4460
#> direct          6 -0.1651 0.0617 20000 -0.3241 -0.0862
#> indirect        6  0.5170 0.0954 20000  0.3772  0.7489
#> 
#> $`7`
#>          interval     est     se     R    2.5%   97.5%
#> total           7  0.3153 0.0439 20000  0.2426  0.4140
#> direct          7 -0.1220 0.0562 20000 -0.2721 -0.0564
#> indirect        7  0.4373 0.0898 20000  0.3110  0.6616
#> 
#> $`8`
#>          interval     est     se     R    2.5%   97.5%
#> total           8  0.2693 0.0442 20000  0.1984  0.3710
#> direct          8 -0.0886 0.0502 20000 -0.2260 -0.0358
#> indirect        8  0.3579 0.0813 20000  0.2488  0.5653
#> 
#> $`9`
#>          interval     est     se     R    2.5%   97.5%
#> total           9  0.2234 0.0438 20000  0.1543  0.3259
#> direct          9 -0.0635 0.0441 20000 -0.1879 -0.0219
#> indirect        9  0.2869 0.0721 20000  0.1929  0.4734
#> 
#> $`10`
#>          interval     est     se     R    2.5%   97.5%
#> total          10  0.1822 0.0431 20000  0.1151  0.2844
#> direct         10 -0.0451 0.0384 20000 -0.1553 -0.0130
#> indirect       10  0.2273 0.0634 20000  0.1469  0.3915
#> 
#> $`11`
#>          interval     est     se     R    2.5%   97.5%
#> total          11  0.1473 0.0422 20000  0.0822  0.2478
#> direct         11 -0.0318 0.0331 20000 -0.1288 -0.0073
#> indirect       11  0.1790 0.0558 20000  0.1097  0.3252
#> 
#> $`12`
#>          interval     est     se     R    2.5%   97.5%
#> total          12  0.1185 0.0408 20000  0.0564  0.2163
#> direct         12 -0.0223 0.0285 20000 -0.1063 -0.0039
#> indirect       12  0.1408 0.0495 20000  0.0801  0.2711
#> 
#> $`13`
#>          interval     est     se     R    2.5%   97.5%
#> total          13  0.0953 0.0391 20000  0.0367  0.1901
#> direct         13 -0.0155 0.0245 20000 -0.0875 -0.0019
#> indirect       13  0.1109 0.0444 20000  0.0572  0.2281
#> 
#> $`14`
#>          interval     est     se     R    2.5%   97.5%
#> total          14  0.0767 0.0370 20000  0.0227  0.1674
#> direct         14 -0.0108 0.0211 20000 -0.0720 -0.0009
#> indirect       14  0.0875 0.0401 20000  0.0401  0.1946
#> 
#> $`15`
#>          interval     est     se     R    2.5%   97.5%
#> total          15  0.0618 0.0346 20000  0.0131  0.1480
#> direct         15 -0.0075 0.0182 20000 -0.0591 -0.0003
#> indirect       15  0.0693 0.0363 20000  0.0275  0.1672
#> 
#> $`16`
#>          interval     est     se     R    2.5%  97.5%
#> total          16  0.0498 0.0320 20000  0.0065 0.1312
#> direct         16 -0.0052 0.0158 20000 -0.0487 0.0000
#> indirect       16  0.0550 0.0330 20000  0.0186 0.1448
#> 
#> $`17`
#>          interval     est     se     R    2.5%  97.5%
#> total          17  0.0402 0.0294 20000  0.0024 0.1162
#> direct         17 -0.0036 0.0138 20000 -0.0400 0.0001
#> indirect       17  0.0438 0.0300 20000  0.0124 0.1260
#> 
#> $`18`
#>          interval     est     se     R    2.5%  97.5%
#> total          18  0.0325 0.0268 20000  0.0001 0.1029
#> direct         18 -0.0025 0.0121 20000 -0.0329 0.0001
#> indirect       18  0.0350 0.0272 20000  0.0081 0.1102
#> 
#> $`19`
#>          interval     est     se     R    2.5%  97.5%
#> total          19  0.0263 0.0243 20000 -0.0012 0.0913
#> direct         19 -0.0017 0.0107 20000 -0.0271 0.0001
#> indirect       19  0.0280 0.0247 20000  0.0053 0.0974
#> 
#> $`20`
#>          interval     est     se     R    2.5%  97.5%
#> total          20  0.0213 0.0219 20000 -0.0018 0.0810
#> direct         20 -0.0012 0.0095 20000 -0.0223 0.0001
#> indirect       20  0.0224 0.0224 20000  0.0034 0.0859
#> 
#> $`21`
#>          interval     est     se     R    2.5%  97.5%
#> total          21  0.0172 0.0197 20000 -0.0018 0.0721
#> direct         21 -0.0008 0.0086 20000 -0.0184 0.0001
#> indirect       21  0.0180 0.0203 20000  0.0021 0.0759
#> 
#> $`22`
#>          interval     est     se     R    2.5%  97.5%
#> total          22  0.0139 0.0177 20000 -0.0015 0.0642
#> direct         22 -0.0005 0.0078 20000 -0.0151 0.0001
#> indirect       22  0.0145 0.0184 20000  0.0014 0.0673
#> 
#> $`23`
#>          interval     est     se     R    2.5%  97.5%
#> total          23  0.0113 0.0159 20000 -0.0013 0.0570
#> direct         23 -0.0004 0.0071 20000 -0.0124 0.0001
#> indirect       23  0.0116 0.0166 20000  0.0008 0.0598
#> 
#> $`24`
#>          interval     est     se     R    2.5%  97.5%
#> total          24  0.0091 0.0143 20000 -0.0010 0.0505
#> direct         24 -0.0003 0.0065 20000 -0.0102 0.0000
#> indirect       24  0.0094 0.0151 20000  0.0005 0.0527
#> 
#> 
#> $delta0
#> $`1`
#>          interval   est     se      z      p    2.5%  97.5%
#> total           1 0.051 0.0434 1.1773 0.2391 -0.0339 0.1360
#> direct          1 0.051 0.0438 1.1639 0.2445 -0.0349 0.1369
#> indirect        1 0.000 0.0009 0.0521 0.9584 -0.0017 0.0018
#> 
#> $`2`
#>          interval     est     se       z      p    2.5%  97.5%
#> total           2  0.0629 0.0531  1.1850 0.2360 -0.0411 0.1669
#> direct          2  0.0637 0.0553  1.1505 0.2499 -0.0448 0.1721
#> indirect        2 -0.0008 0.0027 -0.2919 0.7703 -0.0061 0.0045
#> 
#> $`3`
#>          interval     est     se       z      p    2.5%  97.5%
#> total           3  0.0579 0.0483  1.1987 0.2307 -0.0368 0.1526
#> direct          3  0.0598 0.0524  1.1417 0.2536 -0.0429 0.1625
#> indirect        3 -0.0019 0.0047 -0.4002 0.6890 -0.0112 0.0074
#> 
#> $`4`
#>          interval     est     se       z      p    2.5%  97.5%
#> total           4  0.0477 0.0394  1.2094 0.2265 -0.0296 0.1249
#> direct          4  0.0501 0.0441  1.1364 0.2558 -0.0363 0.1365
#> indirect        4 -0.0024 0.0061 -0.3948 0.6930 -0.0144 0.0096
#> 
#> $`5`
#>          interval     est     se       z      p    2.5%  97.5%
#> total           5  0.0373 0.0309  1.2062 0.2277 -0.0233 0.0979
#> direct          5  0.0394 0.0348  1.1332 0.2571 -0.0288 0.1076
#> indirect        5 -0.0022 0.0068 -0.3168 0.7514 -0.0155 0.0112
#> 
#> $`6`
#>          interval     est     se       z      p    2.5%  97.5%
#> total           6  0.0286 0.0242  1.1802 0.2379 -0.0189 0.0761
#> direct          6  0.0299 0.0264  1.1307 0.2582 -0.0219 0.0817
#> indirect        6 -0.0013 0.0071 -0.1845 0.8537 -0.0153 0.0127
#> 
#> $`7`
#>          interval     est     se       z      p    2.5%  97.5%
#> total           7  0.0219 0.0194  1.1297 0.2586 -0.0161 0.0598
#> direct          7  0.0221 0.0196  1.1271 0.2597 -0.0163 0.0605
#> indirect        7 -0.0002 0.0074 -0.0308 0.9755 -0.0147 0.0142
#> 
#> $`8`
#>          interval    est     se      z      p    2.5%  97.5%
#> total           8 0.0169 0.0159 1.0618 0.2883 -0.0143 0.0480
#> direct          8 0.0161 0.0143 1.1209 0.2623 -0.0120 0.0441
#> indirect        8 0.0008 0.0076 0.1079 0.9141 -0.0141 0.0158
#> 
#> $`9`
#>          interval    est     se      z      p    2.5%  97.5%
#> total           9 0.0132 0.0133 0.9881 0.3231 -0.0130 0.0393
#> direct          9 0.0115 0.0104 1.1103 0.2669 -0.0088 0.0318
#> indirect        9 0.0017 0.0077 0.2154 0.8295 -0.0135 0.0168
#> 
#> $`10`
#>          interval    est     se      z      p    2.5%  97.5%
#> total          10 0.0104 0.0113 0.9180 0.3586 -0.0118 0.0326
#> direct         10 0.0082 0.0075 1.0938 0.2740 -0.0065 0.0228
#> indirect       10 0.0022 0.0076 0.2932 0.7693 -0.0127 0.0172
#> 
#> $`11`
#>          interval    est     se      z      p    2.5%  97.5%
#> total          11 0.0083 0.0097 0.8561 0.3919 -0.0107 0.0273
#> direct         11 0.0058 0.0054 1.0706 0.2844 -0.0048 0.0163
#> indirect       11 0.0025 0.0073 0.3487 0.7273 -0.0118 0.0169
#> 
#> $`12`
#>          interval    est     se      z      p    2.5%  97.5%
#> total          12 0.0067 0.0083 0.8035 0.4217 -0.0096 0.0229
#> direct         12 0.0040 0.0039 1.0403 0.2982 -0.0036 0.0116
#> indirect       12 0.0026 0.0068 0.3882 0.6979 -0.0107 0.0160
#> 
#> $`13`
#>          interval    est     se      z      p    2.5%  97.5%
#> total          13 0.0054 0.0071 0.7589 0.4479 -0.0085 0.0193
#> direct         13 0.0028 0.0028 1.0035 0.3156 -0.0027 0.0083
#> indirect       13 0.0026 0.0062 0.4163 0.6772 -0.0095 0.0147
#> 
#> $`14`
#>          interval    est     se      z      p    2.5%  97.5%
#> total          14 0.0044 0.0060 0.7209 0.4710 -0.0075 0.0162
#> direct         14 0.0020 0.0020 0.9613 0.3364 -0.0020 0.0059
#> indirect       14 0.0024 0.0055 0.4361 0.6628 -0.0084 0.0132
#> 
#> $`15`
#>          interval    est     se      z      p    2.5%  97.5%
#> total          15 0.0035 0.0051 0.6878 0.4916 -0.0065 0.0136
#> direct         15 0.0014 0.0015 0.9154 0.3600 -0.0015 0.0043
#> indirect       15 0.0022 0.0048 0.4496 0.6530 -0.0073 0.0117
#> 
#> $`16`
#>          interval    est     se      z      p    2.5%  97.5%
#> total          16 0.0029 0.0043 0.6584 0.5103 -0.0057 0.0114
#> direct         16 0.0009 0.0011 0.8673 0.3858 -0.0012 0.0030
#> indirect       16 0.0019 0.0042 0.4584 0.6467 -0.0063 0.0102
#> 
#> $`17`
#>          interval    est     se      z      p    2.5%  97.5%
#> total          17 0.0023 0.0037 0.6319 0.5275 -0.0049 0.0095
#> direct         17 0.0006 0.0008 0.8186 0.4130 -0.0009 0.0022
#> indirect       17 0.0017 0.0036 0.4633 0.6432 -0.0054 0.0087
#> 
#> $`18`
#>          interval    est     se      z      p    2.5%  97.5%
#> total          18 0.0019 0.0031 0.6075 0.5435 -0.0042 0.0079
#> direct         18 0.0004 0.0006 0.7708 0.4409 -0.0007 0.0016
#> indirect       18 0.0014 0.0031 0.4651 0.6418 -0.0046 0.0075
#> 
#> $`19`
#>          interval    est     se      z      p    2.5%  97.5%
#> total          19 0.0015 0.0026 0.5848 0.5587 -0.0036 0.0066
#> direct         19 0.0003 0.0004 0.7246 0.4687 -0.0005 0.0011
#> indirect       19 0.0012 0.0026 0.4646 0.6422 -0.0039 0.0063
#> 
#> $`20`
#>          interval    est     se      z      p    2.5%  97.5%
#> total          20 0.0012 0.0022 0.5637 0.5729 -0.0030 0.0055
#> direct         20 0.0002 0.0003 0.6809 0.4960 -0.0004 0.0008
#> indirect       20 0.0010 0.0022 0.4621 0.6440 -0.0033 0.0053
#> 
#> $`21`
#>          interval   est     se      z      p    2.5%  97.5%
#> total          21 1e-03 0.0018 0.5440 0.5865 -0.0026 0.0046
#> direct         21 1e-04 0.0002 0.6399 0.5223 -0.0003 0.0006
#> indirect       21 9e-04 0.0019 0.4581 0.6469 -0.0028 0.0045
#> 
#> $`22`
#>          interval   est     se      z      p    2.5%  97.5%
#> total          22 8e-04 0.0015 0.5254 0.5993 -0.0022 0.0038
#> direct         22 1e-04 0.0002 0.6018 0.5473 -0.0002 0.0004
#> indirect       22 7e-04 0.0016 0.4529 0.6506 -0.0023 0.0038
#> 
#> $`23`
#>          interval   est     se      z      p    2.5%  97.5%
#> total          23 7e-04 0.0013 0.5080 0.6115 -0.0019 0.0032
#> direct         23 1e-04 0.0001 0.5666 0.5710 -0.0002 0.0003
#> indirect       23 6e-04 0.0013 0.4468 0.6551 -0.0020 0.0031
#> 
#> $`24`
#>          interval   est     se      z      p    2.5%  97.5%
#> total          24 5e-04 0.0011 0.4916 0.6230 -0.0016 0.0026
#> direct         24 0e+00 0.0001 0.5342 0.5932 -0.0001 0.0002
#> indirect       24 5e-04 0.0011 0.4400 0.6600 -0.0017 0.0026
#> 
#> 
#> $mc0
#> $`1`
#>          interval   est     se     R    2.5%  97.5%
#> total           1 0.051 0.0439 20000 -0.0341 0.1373
#> direct          1 0.051 0.0444 20000 -0.0339 0.1400
#> indirect        1 0.000 0.0015 20000 -0.0042 0.0023
#> 
#> $`2`
#>          interval     est     se     R    2.5%  97.5%
#> total           2  0.0629 0.0539 20000 -0.0429 0.1691
#> direct          2  0.0637 0.0565 20000 -0.0420 0.1796
#> indirect        2 -0.0008 0.0046 20000 -0.0145 0.0040
#> 
#> $`3`
#>          interval     est     se     R    2.5%  97.5%
#> total           3  0.0579 0.0494 20000 -0.0403 0.1540
#> direct          3  0.0598 0.0541 20000 -0.0395 0.1736
#> indirect        3 -0.0019 0.0077 20000 -0.0253 0.0058
#> 
#> $`4`
#>          interval     est     se     R    2.5%  97.5%
#> total           4  0.0477 0.0407 20000 -0.0338 0.1260
#> direct          4  0.0501 0.0464 20000 -0.0337 0.1500
#> indirect        4 -0.0024 0.0101 20000 -0.0321 0.0091
#> 
#> $`5`
#>          interval     est     se     R    2.5%  97.5%
#> total           5  0.0373 0.0323 20000 -0.0271 0.1005
#> direct          5  0.0394 0.0376 20000 -0.0275 0.1230
#> indirect        5 -0.0022 0.0114 20000 -0.0344 0.0136
#> 
#> $`6`
#>          interval     est     se     R    2.5%  97.5%
#> total           6  0.0286 0.0256 20000 -0.0209 0.0817
#> direct          6  0.0299 0.0296 20000 -0.0218 0.0976
#> indirect        6 -0.0013 0.0119 20000 -0.0327 0.0187
#> 
#> $`7`
#>          interval     est     se     R    2.5%  97.5%
#> total           7  0.0219 0.0208 20000 -0.0150 0.0679
#> direct          7  0.0221 0.0230 20000 -0.0171 0.0760
#> indirect        7 -0.0002 0.0121 20000 -0.0290 0.0237
#> 
#> $`8`
#>          interval    est     se     R    2.5%  97.5%
#> total           8 0.0169 0.0173 20000 -0.0107 0.0583
#> direct          8 0.0161 0.0177 20000 -0.0134 0.0584
#> indirect        8 0.0008 0.0120 20000 -0.0246 0.0275
#> 
#> $`9`
#>          interval    est     se     R    2.5%  97.5%
#> total           9 0.0132 0.0148 20000 -0.0077 0.0507
#> direct          9 0.0115 0.0137 20000 -0.0105 0.0452
#> indirect        9 0.0017 0.0116 20000 -0.0202 0.0297
#> 
#> $`10`
#>          interval    est     se     R    2.5%  97.5%
#> total          10 0.0104 0.0129 20000 -0.0053 0.0448
#> direct         10 0.0082 0.0107 20000 -0.0081 0.0352
#> indirect       10 0.0022 0.0111 20000 -0.0163 0.0308
#> 
#> $`11`
#>          interval    est     se     R    2.5%  97.5%
#> total          11 0.0083 0.0113 20000 -0.0037 0.0398
#> direct         11 0.0058 0.0084 20000 -0.0062 0.0273
#> indirect       11 0.0025 0.0105 20000 -0.0132 0.0306
#> 
#> $`12`
#>          interval    est     se     R    2.5%  97.5%
#> total          12 0.0067 0.0100 20000 -0.0026 0.0354
#> direct         12 0.0040 0.0067 20000 -0.0048 0.0214
#> indirect       12 0.0026 0.0097 20000 -0.0103 0.0293
#> 
#> $`13`
#>          interval    est     se     R    2.5%  97.5%
#> total          13 0.0054 0.0089 20000 -0.0018 0.0314
#> direct         13 0.0028 0.0054 20000 -0.0037 0.0168
#> indirect       13 0.0026 0.0089 20000 -0.0082 0.0274
#> 
#> $`14`
#>          interval    est     se     R    2.5%  97.5%
#> total          14 0.0044 0.0079 20000 -0.0013 0.0278
#> direct         14 0.0020 0.0044 20000 -0.0028 0.0133
#> indirect       14 0.0024 0.0081 20000 -0.0066 0.0254
#> 
#> $`15`
#>          interval    est     se     R    2.5%  97.5%
#> total          15 0.0035 0.0070 20000 -0.0009 0.0246
#> direct         15 0.0014 0.0036 20000 -0.0022 0.0104
#> indirect       15 0.0022 0.0073 20000 -0.0052 0.0232
#> 
#> $`16`
#>          interval    est     se     R    2.5%  97.5%
#> total          16 0.0029 0.0062 20000 -0.0006 0.0219
#> direct         16 0.0009 0.0030 20000 -0.0017 0.0082
#> indirect       16 0.0019 0.0065 20000 -0.0042 0.0209
#> 
#> $`17`
#>          interval    est     se     R    2.5%  97.5%
#> total          17 0.0023 0.0055 20000 -0.0004 0.0194
#> direct         17 0.0006 0.0025 20000 -0.0013 0.0064
#> indirect       17 0.0017 0.0059 20000 -0.0034 0.0187
#> 
#> $`18`
#>          interval    est     se     R    2.5%  97.5%
#> total          18 0.0019 0.0049 20000 -0.0003 0.0173
#> direct         18 0.0004 0.0021 20000 -0.0010 0.0050
#> indirect       18 0.0014 0.0052 20000 -0.0027 0.0167
#> 
#> $`19`
#>          interval    est     se     R    2.5%  97.5%
#> total          19 0.0015 0.0044 20000 -0.0002 0.0153
#> direct         19 0.0003 0.0018 20000 -0.0008 0.0040
#> indirect       19 0.0012 0.0047 20000 -0.0022 0.0150
#> 
#> $`20`
#>          interval    est     se     R    2.5%  97.5%
#> total          20 0.0012 0.0039 20000 -0.0002 0.0137
#> direct         20 0.0002 0.0016 20000 -0.0006 0.0032
#> indirect       20 0.0010 0.0042 20000 -0.0018 0.0135
#> 
#> $`21`
#>          interval   est     se     R    2.5%  97.5%
#> total          21 1e-03 0.0035 20000 -0.0001 0.0121
#> direct         21 1e-04 0.0014 20000 -0.0005 0.0025
#> indirect       21 9e-04 0.0038 20000 -0.0014 0.0120
#> 
#> $`22`
#>          interval   est     se     R    2.5%  97.5%
#> total          22 8e-04 0.0032 20000 -0.0001 0.0107
#> direct         22 1e-04 0.0012 20000 -0.0004 0.0020
#> indirect       22 7e-04 0.0034 20000 -0.0011 0.0107
#> 
#> $`23`
#>          interval   est     se     R   2.5%  97.5%
#> total          23 7e-04 0.0029 20000 -1e-04 0.0096
#> direct         23 1e-04 0.0011 20000 -3e-04 0.0016
#> indirect       23 6e-04 0.0030 20000 -9e-04 0.0095
#> 
#> $`24`
#>          interval   est     se     R   2.5%  97.5%
#> total          24 5e-04 0.0026 20000 -1e-04 0.0085
#> direct         24 0e+00 0.0010 20000 -2e-04 0.0013
#> indirect       24 5e-04 0.0027 20000 -7e-04 0.0085
#> 
#>