Parametric Bootstrap (Illustration)

The function generates simulated datasets based on a fitted model and refits the model to each generated dataset using the dynr package.

Usage

IllustrationBootPara(
  fit,
  path,
  prefix,
  taskid,
  B = 1000L,
  ncores = NULL,
  seed = NULL
)

Arguments

fit

R object. Fitted CT-VAR model.

path

Path to a directory to store bootstrap samples and estimates.

prefix

Character string. Prefix used for the file names for the bootstrap samples and estimates.

taskid

Positive integer. Task ID.

B

Positive integer. Number of bootstrap samples.

ncores

Positive integer. Number of cores to use.

seed

Integer. Random seed.

See also

Examples

if (FALSE) { # \dontrun{
library(dynr)
sim <- IllustrationGenData(seed = 42)
data <- IllustrationPrepData(sim)
fit <- IllustrationFitDynr(data)
summary(fit)
IllustrationBootPara(
  fit = fit,
  path = getwd(),
  prefix = "pb",
  taskid = 1,
  B = 1000L,
  seed = 42
)
} # }