Generate a Sampling Distribution of Drift Matrices and Process Noise Covariance Matrices (Illustration)

The function generates a sampling distribution of drift matrices and process noise covariance matrices using te Monte Carlo method.

Usage

IllustrationMCPhiSigma(fit, R = 20000L, seed = NULL)

Arguments

fit

R object. Fitted CT-VAR model.

R

Positive integer. Number of Monte Carlo replications.

seed

Integer. Random seed.

See also

Examples

if (FALSE) { # \dontrun{
library(dynr)
sim <- IllustrationGenData(seed = 42)
data <- IllustrationPrepData(sim)
fit <- IllustrationFitDynr(data)
IllustrationMCPhiSigma(fit, seed = 42)
} # }