Generate Confidence Intervals for the Total, Direct, and Indirect Effects

The function generates confidence intervals for the total, direct, amd indirect effects using the cTMed package.

Usage

CI(x, delta_t = c(5, 10, 15, 20), R = 20000L)

Arguments

x

R object. Output of the FitCtsem() or FitDynr() functions.

delta_t

Numeric vector. Vector of time-intervals to use.

R

Positive integer. Number of Monte Carlo replications.

See also

Examples

# dynr
library(dynr)
#> Loading required package: ggplot2
data <- Data(repid = 1, n = 50)
fit <- FitDynr(x = data)
#> [1] "Get ready!!!!"
#> using C compiler: ‘gcc (Ubuntu 11.4.0-1ubuntu1~22.04) 11.4.0’
#> Optimization function called.
#> Starting Hessian calculation ...
#> Finished Hessian calculation.
#> Original exit flag:  3 
#> Modified exit flag:  3 
#> Optimization terminated successfully: ftol_rel or ftol_abs was reached. 
#> Original fitted parameters:  -0.2605694 -0.0743651 0.08194417 0.6720199 
#> -0.460244 -0.02385756 -0.4523976 0.7365811 -0.6910508 -1.475439 0.2015311 
#> -0.2468988 -2.729321 0.2947053 -2.877131 -1.612013 -1.617904 -1.576812 
#> -0.1284087 0.06218073 0.2874795 -0.175533 0.2349609 -0.1969349 
#> 
#> Transformed fitted parameters:  -0.2605694 -0.0743651 0.08194417 0.6720199 
#> -0.460244 -0.02385756 -0.4523976 0.7365811 -0.6910508 0.2286783 0.04608579 
#> -0.0564604 0.07455129 0.007854991 0.07590425 0.1994856 0.198314 0.2066329 
#> 0.8794939 0.05468757 0.2528364 0.8424102 0.212856 0.9402482 
#> 
#> Doing end processing
#> Successful trial
#> Total Time: 3.825781 
#> Backend Time: 3.817819 
CI(fit)
#> $type
#> [1] "dynr"
#> 
#> $delta_t
#> [1]  5 10 15 20
#> 
#> $delta
#> $`5`
#>          interval     est     se       z     p    2.5%   97.5%
#> total           5  0.3629 0.0390  9.3003 0e+00  0.2865  0.4394
#> direct          5 -0.2177 0.0607 -3.5847 3e-04 -0.3368 -0.0987
#> indirect        5  0.5807 0.0912  6.3681 0e+00  0.4020  0.7594
#> 
#> $`10`
#>          interval     est     se       z      p    2.5%  97.5%
#> total          10  0.1822 0.0414  4.4022 0.0000  0.1011 0.2633
#> direct         10 -0.0451 0.0278 -1.6231 0.1046 -0.0996 0.0094
#> indirect       10  0.2273 0.0545  4.1701 0.0000  0.1205 0.3341
#> 
#> $`15`
#>          interval     est     se       z      p    2.5%  97.5%
#> total          15  0.0618 0.0315  1.9637 0.0496  0.0001 0.1235
#> direct         15 -0.0075 0.0081 -0.9233 0.3558 -0.0233 0.0084
#> indirect       15  0.0693 0.0305  2.2693 0.0232  0.0094 0.1291
#> 
#> $`20`
#>          interval     est     se       z      p    2.5%  97.5%
#> total          20  0.0213 0.0179  1.1885 0.2347 -0.0138 0.0564
#> direct         20 -0.0012 0.0019 -0.6086 0.5428 -0.0049 0.0026
#> indirect       20  0.0224 0.0172  1.3025 0.1928 -0.0113 0.0562
#> 
#> 
#> $mc
#> $`5`
#>          interval     est     se     R    2.5%   97.5%
#> total           5  0.3629 0.0393 20000  0.2935  0.4472
#> direct          5 -0.2177 0.0655 20000 -0.3797 -0.1263
#> indirect        5  0.5807 0.0957 20000  0.4337  0.8067
#> 
#> $`10`
#>          interval     est     se     R    2.5%   97.5%
#> total          10  0.1822 0.0431 20000  0.1151  0.2844
#> direct         10 -0.0451 0.0384 20000 -0.1553 -0.0130
#> indirect       10  0.2273 0.0634 20000  0.1469  0.3915
#> 
#> $`15`
#>          interval     est     se     R    2.5%   97.5%
#> total          15  0.0618 0.0346 20000  0.0131  0.1480
#> direct         15 -0.0075 0.0182 20000 -0.0591 -0.0003
#> indirect       15  0.0693 0.0363 20000  0.0275  0.1672
#> 
#> $`20`
#>          interval     est     se     R    2.5%  97.5%
#> total          20  0.0213 0.0219 20000 -0.0018 0.0810
#> direct         20 -0.0012 0.0095 20000 -0.0223 0.0001
#> indirect       20  0.0224 0.0224 20000  0.0034 0.0859
#> 
#> 
#> $posterior
#> NULL
#> 

# ctsem
library(ctsem)
#> Loading required package: Rcpp
#> ctsem also changes in time, for manual run ctDocs(), for blog see https://cdriver.netlify.app/, for citation info run citation('ctsem'), for original OpenMx functionality install.packages('ctsemOMX'), and for discussion https://github.com/cdriveraus/ctsem/discussions
fit <- FitCtsem(x = data)
#> System dimensions inferred from LAMBDA
#> Maximum likelihood estimation requested
#> Free T0VAR parameters as well as indvarying T0MEANS -- fixing T0VAR pars to diag matrix of 1e-3
#> Using 1/4 logical CPU cores
#> 1st pass optimization (carefulfit)...
#> Progress est. = 0%, LPchange = 260, Iter = 155, LP = -13369.472185547, worseCount = 0           
#> Progress est. = 4.54%, LPchange = 160, Iter = 160, LP = -13361.5138212261, worseCount = 0           
#> Progress est. = 10.67%, LPchange = 88, Iter = 165, LP = -13361.5138212261, worseCount = 0           
#> Progress est. = 16.42%, LPchange = 49, Iter = 170, LP = -13361.5138212261, worseCount = 1           
#> Progress est. = 17.74%, LPchange = 43, Iter = 175, LP = -13361.5138212261, worseCount = 2           
#> Progress est. = 18.44%, LPchange = 40, Iter = 180, LP = -13361.5138212261, worseCount = 3           
#> Progress est. = 18.99%, LPchange = 38, Iter = 185, LP = -13361.5138212261, worseCount = 4           
#> Progress est. = 20.22%, LPchange = 33, Iter = 190, LP = -13361.5138212261, worseCount = 5           
#> Progress est. = 21.97%, LPchange = 28, Iter = 195, LP = -13361.5138212261, worseCount = 6           
#> Progress est. = 23.65%, LPchange = 24, Iter = 200, LP = -13361.5138212261, worseCount = 7           
#> Progress est. = 33.19%, LPchange = 8.9, Iter = 205, LP = -13361.5138212261, worseCount = 8           
#> Progress est. = 33.19%, LPchange = 8.9, Iter = 210, LP = -13361.5138212261, worseCount = 9           
#> Progress est. = 24.85%, LPchange = 21, Iter = 215, LP = -13361.5138212261, worseCount = 10           
#> Progress est. = 22.78%, LPchange = 26, Iter = 220, LP = -13361.5138212261, worseCount = 11           
#> Progress est. = 21.83%, LPchange = 28, Iter = 225, LP = -13361.5138212261, worseCount = 12           
#> Progress est. = 22.64%, LPchange = 26, Iter = 230, LP = -13361.5138212261, worseCount = 13           
#> Progress est. = 23.85%, LPchange = 23, Iter = 235, LP = -13361.5138212261, worseCount = 14           
#> Progress est. = 26.39%, LPchange = 18, Iter = 240, LP = -13361.5138212261, worseCount = 15           
#> Progress est. = 30.06%, LPchange = 12, Iter = 245, LP = -13361.5138212261, worseCount = 16           
#> Progress est. = 30.06%, LPchange = 12, Iter = 250, LP = -13361.5138212261, worseCount = 17           
#> Progress est. = 28.12%, LPchange = 15, Iter = 255, LP = -13361.5138212261, worseCount = 18           
#> Progress est. = 26.03%, LPchange = 18, Iter = 260, LP = -13361.5138212261, worseCount = 19           
#> Progress est. = 26.98%, LPchange = 17, Iter = 265, LP = -13361.5138212261, worseCount = 20           
#> 
#> Converged -- count of non-improving iterations exceeded
#> Optimizing...
#> Progress est. = 0%, LPchange = 1.6, Iter = 31, LP = -13224.3227759783
#> Progress est. = 0%, LPchange = 1.7, Iter = 32, LP = -13217.5610891065
#> Progress est. = 0.19%, LPchange = 1.7, Iter = 33, LP = -13213.4785620506
#> Progress est. = 0.78%, LPchange = 1.5, Iter = 34, LP = -13213.4785620506
#> Progress est. = 0%, LPchange = 1.9, Iter = 35, LP = -13201.1422381571
#> Progress est. = 0%, LPchange = 2, Iter = 36, LP = -13198.4826020646
#> Progress est. = 0.04%, LPchange = 1.9, Iter = 37, LP = -13198.4278565559
#> Progress est. = 0%, LPchange = 2.2, Iter = 38, LP = -13190.3339043863
#> Progress est. = 0%, LPchange = 2.4, Iter = 39, LP = -13183.0895604061
#> Progress est. = 0%, LPchange = 2.5, Iter = 40, LP = -13180.4665165906
#> Progress est. = 0.06%, LPchange = 2.5, Iter = 41, LP = -13180.3085877415
#> Progress est. = 0%, LPchange = 2.8, Iter = 42, LP = -13168.9435739441
#> Progress est. = 0.05%, LPchange = 2.8, Iter = 43, LP = -13168.9435739441
#> Progress est. = 0.06%, LPchange = 2.8, Iter = 44, LP = -13168.9435739441
#> Progress est. = 0.12%, LPchange = 2.8, Iter = 45, LP = -13168.9435739441
#> Progress est. = 0.16%, LPchange = 2.8, Iter = 46, LP = -13168.9435739441
#> Progress est. = 0.19%, LPchange = 2.8, Iter = 47, LP = -13168.9435739441
#> Progress est. = 0.22%, LPchange = 2.7, Iter = 48, LP = -13168.9435739441
#> Progress est. = 0.24%, LPchange = 2.7, Iter = 49, LP = -13168.9435739441
#> Progress est. = 0.28%, LPchange = 2.7, Iter = 50, LP = -13168.9435739441
#> Progress est. = 0.31%, LPchange = 2.7, Iter = 51, LP = -13168.9435739441
#> Progress est. = 0.35%, LPchange = 2.7, Iter = 52, LP = -13168.9435739441
#> Progress est. = 0.4%, LPchange = 2.7, Iter = 53, LP = -13168.9435739441
#> Progress est. = 0.48%, LPchange = 2.7, Iter = 54, LP = -13168.9435739441
#> Progress est. = 0.59%, LPchange = 2.6, Iter = 55, LP = -13168.9435739441
#> Progress est. = 0.73%, LPchange = 2.6, Iter = 56, LP = -13168.9435739441
#> Progress est. = 0.91%, LPchange = 2.5, Iter = 57, LP = -13168.9435739441
#> Progress est. = 1.2%, LPchange = 2.4, Iter = 58, LP = -13168.9435739441
#> Progress est. = 1.61%, LPchange = 2.3, Iter = 59, LP = -13168.2930894792
#> Progress est. = 2.34%, LPchange = 2.1, Iter = 60, LP = -13168.2930894792
#> Progress est. = 2.41%, LPchange = 2.1, Iter = 61, LP = -13161.7963937833
#> Progress est. = 3.32%, LPchange = 1.9, Iter = 62, LP = -13161.7963937833
#> Progress est. = 3.93%, LPchange = 1.7, Iter = 63, LP = -13161.7963937833
#> Progress est. = 3.93%, LPchange = 1.7, Iter = 64, LP = -13161.7963937833
#> Progress est. = 6.08%, LPchange = 1.3, Iter = 65, LP = -13161.7240769137
#> Progress est. = 6.64%, LPchange = 1.2, Iter = 66, LP = -13161.7240769137
#> Progress est. = 6.45%, LPchange = 1.3, Iter = 67, LP = -13160.7819878742
#> Progress est. = 8.38%, LPchange = 0.99, Iter = 68, LP = -13160.7819878742
#> Progress est. = 10.52%, LPchange = 0.75, Iter = 69, LP = -13160.5050271523
#> Progress est. = 11.51%, LPchange = 0.67, Iter = 70, LP = -13160.5050271523
#> Progress est. = 11.48%, LPchange = 0.67, Iter = 71, LP = -13160.282619818
#> Progress est. = 18.16%, LPchange = 0.29, Iter = 72, LP = -13160.282619818
#> Progress est. = 18.11%, LPchange = 0.29, Iter = 73, LP = -13160.2337456041
#> Progress est. = 18.11%, LPchange = 0.29, Iter = 74, LP = -13160.2337456041
#> Progress est. = 18.08%, LPchange = 0.29, Iter = 75, LP = -13160.1993401639
#> Progress est. = 18.08%, LPchange = 0.29, Iter = 76, LP = -13160.1993401639
#> Progress est. = 18.03%, LPchange = 0.29, Iter = 77, LP = -13160.1363825578
#> Progress est. = 18.03%, LPchange = 0.29, Iter = 78, LP = -13160.1363825578
#> Progress est. = 17.9%, LPchange = 0.3, Iter = 79, LP = -13159.9922056438
#> Progress est. = 17.9%, LPchange = 0.3, Iter = 80, LP = -13159.9922056438
#> Progress est. = 17.9%, LPchange = 0.3, Iter = 81, LP = -13159.9922056438
#> Progress est. = 17.9%, LPchange = 0.3, Iter = 82, LP = -13159.9922056438
#> Progress est. = 17.89%, LPchange = 0.3, Iter = 83, LP = -13159.985630316
#> Progress est. = 17.89%, LPchange = 0.3, Iter = 84, LP = -13159.985630316
#> Progress est. = 17.87%, LPchange = 0.3, Iter = 85, LP = -13159.9614487546
#> Progress est. = 17.87%, LPchange = 0.3, Iter = 86, LP = -13159.9570742485
#> Progress est. = 17.85%, LPchange = 0.3, Iter = 87, LP = -13159.9358253586
#> Progress est. = 17.83%, LPchange = 0.3, Iter = 88, LP = -13159.9157300208
#> Progress est. = 18.42%, LPchange = 0.28, Iter = 89, LP = -13159.9157300208
#> Progress est. = 18.41%, LPchange = 0.28, Iter = 90, LP = -13159.8963188403
#> Progress est. = 30.25%, LPchange = 0.063, Iter = 91, LP = -13159.8963188403
#> Progress est. = 30.2%, LPchange = 0.064, Iter = 92, LP = -13159.8848738294
#> Progress est. = 30.2%, LPchange = 0.064, Iter = 93, LP = -13159.8848738294
#> Progress est. = 30.17%, LPchange = 0.064, Iter = 94, LP = -13159.8769632149
#> Progress est. = 30.46%, LPchange = 0.062, Iter = 95, LP = -13159.8746271477
#> Progress est. = 30.44%, LPchange = 0.062, Iter = 96, LP = -13159.8695661445
#> Progress est. = 36.04%, LPchange = 0.031, Iter = 97, LP = -13159.8632976016
#> Progress est. = 36.02%, LPchange = 0.031, Iter = 98, LP = -13159.8607682837
#> Progress est. = 38.8%, LPchange = 0.022, Iter = 99, LP = -13159.8555922399
#> Progress est. = 38.76%, LPchange = 0.022, Iter = 100, LP = -13159.8518297808
#> Progress est. = 42.01%, LPchange = 0.014, Iter = 101, LP = -13159.8482045035
#> Progress est. = 41.95%, LPchange = 0.015, Iter = 102, LP = -13159.8453042184
#> Progress est. = 42.83%, LPchange = 0.013, Iter = 103, LP = -13159.841799358
#> Progress est. = 42.77%, LPchange = 0.013, Iter = 104, LP = -13159.8390760071
#> Progress est. = 43.45%, LPchange = 0.012, Iter = 105, LP = -13159.8367315918
#> Progress est. = 43.4%, LPchange = 0.012, Iter = 106, LP = -13159.8344525862
#> Progress est. = 44.85%, LPchange = 0.01, Iter = 107, LP = -13159.8324780865
#> Progress est. = 44.81%, LPchange = 0.01, Iter = 108, LP = -13159.831018778
#> Progress est. = 49.85%, LPchange = 0.0054, Iter = 109, LP = -13159.8298026021
#> Progress est. = 49.81%, LPchange = 0.0054, Iter = 110, LP = -13159.8290687747
#> Progress est. = 49.78%, LPchange = 0.0055, Iter = 111, LP = -13159.8285365708
#> Progress est. = 49.78%, LPchange = 0.0055, Iter = 112, LP = -13159.8283624214
#> Progress est. = 50.1%, LPchange = 0.0052, Iter = 113, LP = -13159.8283624214
#> Progress est. = 50.1%, LPchange = 0.0052, Iter = 114, LP = -13159.8283624214
#> Progress est. = 51.43%, LPchange = 0.0044, Iter = 115, LP = -13159.8283624214
#> Progress est. = 51.7%, LPchange = 0.0043, Iter = 116, LP = -13159.8283624214
#> Progress est. = 53.14%, LPchange = 0.0036, Iter = 117, LP = -13159.8283624214
#> Progress est. = 54.79%, LPchange = 0.0029, Iter = 118, LP = -13159.8283624214
#> Progress est. = 54.79%, LPchange = 0.0029, Iter = 119, LP = -13159.8283624214
#> Progress est. = 56.79%, LPchange = 0.0023, Iter = 120, LP = -13159.8283624214
#> Progress est. = 56.79%, LPchange = 0.0023, Iter = 121, LP = -13159.8283624214
#> Progress est. = 58.26%, LPchange = 0.0019, Iter = 122, LP = -13159.8283624214
#> Progress est. = 58.26%, LPchange = 0.0019, Iter = 123, LP = -13159.8283624214
#> Progress est. = 59.46%, LPchange = 0.0016, Iter = 124, LP = -13159.8283624214
#> Progress est. = 59.85%, LPchange = 0.0015, Iter = 125, LP = -13159.8283624214
#> Progress est. = 60.78%, LPchange = 0.0014, Iter = 126, LP = -13159.8283624214
#> Progress est. = 62.09%, LPchange = 0.0012, Iter = 127, LP = -13159.8283624214
#> Progress est. = 62.69%, LPchange = 0.0011, Iter = 128, LP = -13159.8283624214
#> Progress est. = 64.08%, LPchange = 0.00091, Iter = 129, LP = -13159.8283624214
#> Progress est. = 65.26%, LPchange = 0.00078, Iter = 130, LP = -13159.8283624214
#> Progress est. = 66.6%, LPchange = 0.00066, Iter = 131, LP = -13159.8283624214
#> Progress est. = 67.84%, LPchange = 0.00057, Iter = 132, LP = -13159.8283276855
#> Progress est. = 69.68%, LPchange = 0.00045, Iter = 133, LP = -13159.8283276855
#> Progress est. = 71.07%, LPchange = 0.00038, Iter = 134, LP = -13159.8277582101
#> Progress est. = 72.92%, LPchange = 3e-04, Iter = 135, LP = -13159.8277582101
#> Progress est. = 74.83%, LPchange = 0.00024, Iter = 136, LP = -13159.8273895533
#> Progress est. = 77.44%, LPchange = 0.00017, Iter = 137, LP = -13159.8273895533
#> Progress est. = 80.13%, LPchange = 0.00012, Iter = 138, LP = -13159.8273895533
#> Progress est. = 83.39%, LPchange = 8e-05, Iter = 139, LP = -13159.8273895533
#> Progress est. = 86.23%, LPchange = 5.6e-05, Iter = 140, LP = -13159.8273797395
#> Progress est. = 89.25%, LPchange = 3.9e-05, Iter = 141, LP = -13159.8273797395
#> Progress est. = 90.07%, LPchange = 3.5e-05, Iter = 142, LP = -13159.8273198163
#> Progress est. = 90.07%, LPchange = 3.5e-05, Iter = 143, LP = -13159.8273198163
#> Progress est. = 90.07%, LPchange = 3.5e-05, Iter = 144, LP = -13159.8273198163
#> Progress est. = 90.07%, LPchange = 3.5e-05, Iter = 145, LP = -13159.8273198163
#> Progress est. = 89.76%, LPchange = 3.6e-05, Iter = 146, LP = -13159.8272782677
#> Progress est. = 89.76%, LPchange = 3.6e-05, Iter = 147, LP = -13159.8272782677
#> Progress est. = 89.73%, LPchange = 3.6e-05, Iter = 148, LP = -13159.827274416
#> Progress est. = 89.73%, LPchange = 3.6e-05, Iter = 149, LP = -13159.827274416
#> Progress est. = 89.66%, LPchange = 3.7e-05, Iter = 150, LP = -13159.8272640787
#> Progress est. = 89.66%, LPchange = 3.7e-05, Iter = 151, LP = -13159.8272640787
#> Progress est. = 89.63%, LPchange = 3.7e-05, Iter = 152, LP = -13159.8272602541
#> Progress est. = 89.63%, LPchange = 3.7e-05, Iter = 153, LP = -13159.8272602541
#> Progress est. = 89.62%, LPchange = 3.7e-05, Iter = 154, LP = -13159.8272583441
#> Progress est. = 89.62%, LPchange = 3.7e-05, Iter = 155, LP = -13159.8272583441
#> Progress est. = 89.62%, LPchange = 3.7e-05, Iter = 156, LP = -13159.8272583441
#> Progress est. = 89.62%, LPchange = 3.7e-05, Iter = 157, LP = -13159.8272583441
#> Progress est. = 89.6%, LPchange = 3.7e-05, Iter = 158, LP = -13159.8272562481
#> Progress est. = 89.6%, LPchange = 3.7e-05, Iter = 159, LP = -13159.8272562481
#> Progress est. = 89.59%, LPchange = 3.7e-05, Iter = 160, LP = -13159.8272547056
#> Progress est. = 89.59%, LPchange = 3.7e-05, Iter = 161, LP = -13159.8272547056
#> Progress est. = 89.82%, LPchange = 3.6e-05, Iter = 162, LP = -13159.8272514732
#> Progress est. = 89.82%, LPchange = 3.6e-05, Iter = 163, LP = -13159.8272514049
#> Progress est. = 95.78%, LPchange = 1.7e-05, Iter = 164, LP = -13159.8272486963
#> Progress est. = 95.78%, LPchange = 1.7e-05, Iter = 165, LP = -13159.8272484946
#> Progress est. = 100%, LPchange = 4.8e-06, Iter = 166, LP = -13159.8272467162
#> 
#> Converged -- lp change within tol(1e-05)
#> Finishing optimization...
#> Estimating Hessian
#> Estimating Hessian, par 1,2%
#> Estimating Hessian, par 2,4%
#> Estimating Hessian, par 3,6%
#> Estimating Hessian, par 4,8%
#> Estimating Hessian, par 5,10%
#> Estimating Hessian, par 6,12%
#> Estimating Hessian, par 7,14%
#> Estimating Hessian, par 8,16%
#> Estimating Hessian, par 9,18%
#> Estimating Hessian, par 10,20%
#> Estimating Hessian, par 11,22%
#> Estimating Hessian, par 12,25%
#> Estimating Hessian, par 13,27%
#> Estimating Hessian, par 14,29%
#> Estimating Hessian, par 15,31%
#> Estimating Hessian, par 16,33%
#> Estimating Hessian, par 17,35%
#> Estimating Hessian, par 18,37%
#> Estimating Hessian, par 19,39%
#> Estimating Hessian, par 20,41%
#> Estimating Hessian, par 21,43%
#> Estimating Hessian, par 22,45%
#> Estimating Hessian, par 23,47%
#> Estimating Hessian, par 24,50%
#> Estimating Hessian, par 1,52%
#> Estimating Hessian, par 2,54%
#> Estimating Hessian, par 3,56%
#> Estimating Hessian, par 4,58%
#> Estimating Hessian, par 5,60%
#> Estimating Hessian, par 6,62%
#> Estimating Hessian, par 7,64%
#> Estimating Hessian, par 8,66%
#> Estimating Hessian, par 9,68%
#> Estimating Hessian, par 10,70%
#> Estimating Hessian, par 11,72%
#> Estimating Hessian, par 12,75%
#> Estimating Hessian, par 13,77%
#> Estimating Hessian, par 14,79%
#> Estimating Hessian, par 15,81%
#> Estimating Hessian, par 16,83%
#> Estimating Hessian, par 17,85%
#> Estimating Hessian, par 18,87%
#> Estimating Hessian, par 19,89%
#> Estimating Hessian, par 20,91%
#> Estimating Hessian, par 21,93%
#> Estimating Hessian, par 22,95%
#> Estimating Hessian, par 23,97%
#> Estimating Hessian, par 24,100%
#> 
#> Computing posterior with 1000 samples
CI(fit)
#> $type
#> [1] "ctsem"
#> 
#> $delta_t
#> [1]  5 10 15 20
#> 
#> $delta
#> $`5`
#>          interval est se       z      p 2.5% 97.5%
#> total           5   0  0  1.3734 0.1696    0     0
#> direct          5   0  0 -0.6601 0.5092    0     0
#> indirect        5   0  0  1.3734 0.1696    0     0
#> 
#> $`10`
#>          interval est se       z      p 2.5% 97.5%
#> total          10   0  0  0.6861 0.4927    0     0
#> direct         10   0  0 -0.3302 0.7413    0     0
#> indirect       10   0  0  0.6861 0.4927    0     0
#> 
#> $`15`
#>          interval est se      z      p 2.5% 97.5%
#> total          15   0  0 0.4569 0.6478    0     0
#> direct         15   0  0 0.3136 0.7538    0     0
#> indirect       15   0  0 0.4569 0.6478    0     0
#> 
#> $`20`
#>          interval est se       z     p 2.5% 97.5%
#> total          20   0  0  0.3424 0.732    0     0
#> direct         20   0  0 -0.1650 0.869    0     0
#> indirect       20   0  0  0.3424 0.732    0     0
#> 
#> 
#> $mc
#> $`5`
#>          interval est se     R 2.5% 97.5%
#> total           5   0  0 20000    0     0
#> direct          5   0  0 20000    0     0
#> indirect        5   0  0 20000    0     0
#> 
#> $`10`
#>          interval est se     R 2.5% 97.5%
#> total          10   0  0 20000    0     0
#> direct         10   0  0 20000    0     0
#> indirect       10   0  0 20000    0     0
#> 
#> $`15`
#>          interval est se     R 2.5% 97.5%
#> total          15   0  0 20000    0     0
#> direct         15   0  0 20000    0     0
#> indirect       15   0  0 20000    0     0
#> 
#> $`20`
#>          interval est se     R 2.5% 97.5%
#> total          20   0  0 20000    0     0
#> direct         20   0  0 20000    0     0
#> indirect       20   0  0 20000    0     0
#> 
#> 
#> $posterior
#> $`5`
#>          interval est se    R 2.5% 97.5%
#> total           5   0  0 1000    0     0
#> direct          5   0  0 1000    0     0
#> indirect        5   0  0 1000    0     0
#> 
#> $`10`
#>          interval est se    R 2.5% 97.5%
#> total          10   0  0 1000    0     0
#> direct         10   0  0 1000    0     0
#> indirect       10   0  0 1000    0     0
#> 
#> $`15`
#>          interval est se    R 2.5% 97.5%
#> total          15   0  0 1000    0     0
#> direct         15   0  0 1000    0     0
#> indirect       15   0  0 1000    0     0
#> 
#> $`20`
#>          interval est se    R 2.5% 97.5%
#> total          20   0  0 1000    0     0
#> direct         20   0  0 1000    0     0
#> indirect       20   0  0 1000    0     0
#> 
#>