Generate Confidence Intervals for the Total, Direct, and Indirect Effects

The function generates confidence intervals for the total, direct, amd indirect effects using the cTMed package.

Usage

CI(x, delta_t = c(5, 10, 15, 20), R = 20000L)

Arguments

x

R object. Output of the FitDynr() function.

delta_t

Numeric vector. Vector of time-intervals to use.

R

Positive integer. Number of Monte Carlo replications.

See also

Examples

# dynr
library(dynr)
#> Loading required package: ggplot2
data <- Data(repid = 1, n = 50)
fit <- FitDynr(x = data)
#> [1] "Get ready!!!!"
#> using C compiler: ‘gcc (Ubuntu 11.4.0-1ubuntu1~22.04) 11.4.0’
#> Optimization function called.
#> Starting Hessian calculation ...
#> Finished Hessian calculation.
#> Original exit flag:  3 
#> Modified exit flag:  3 
#> Optimization terminated successfully: ftol_rel or ftol_abs was reached. 
#> Original fitted parameters:  -0.2605694 -0.0743651 0.08194417 0.6720199 
#> -0.460244 -0.02385756 -0.4523976 0.7365811 -0.6910508 -1.475439 0.2015311 
#> -0.2468988 -2.729321 0.2947053 -2.877131 -1.612013 -1.617904 -1.576812 
#> -0.1284087 0.06218073 0.2874795 -0.175533 0.2349609 -0.1969349 
#> 
#> Transformed fitted parameters:  -0.2605694 -0.0743651 0.08194417 0.6720199 
#> -0.460244 -0.02385756 -0.4523976 0.7365811 -0.6910508 0.2286783 0.04608579 
#> -0.0564604 0.07455129 0.007854991 0.07590425 0.1994856 0.198314 0.2066329 
#> 0.8794939 0.05468757 0.2528364 0.8424102 0.212856 0.9402482 
#> 
#> Doing end processing
#> Successful trial
#> Total Time: 3.840105 
#> Backend Time: 3.831656 
CI(fit)
#> $type
#> [1] "dynr"
#> 
#> $delta_t
#> [1]  5 10 15 20
#> 
#> $delta
#> $`5`
#>          interval     est     se       z     p    2.5%   97.5%
#> total           5  0.3629 0.0390  9.3003 0e+00  0.2865  0.4394
#> direct          5 -0.2177 0.0607 -3.5847 3e-04 -0.3368 -0.0987
#> indirect        5  0.5807 0.0912  6.3681 0e+00  0.4020  0.7594
#> 
#> $`10`
#>          interval     est     se       z      p    2.5%  97.5%
#> total          10  0.1822 0.0414  4.4022 0.0000  0.1011 0.2633
#> direct         10 -0.0451 0.0278 -1.6231 0.1046 -0.0996 0.0094
#> indirect       10  0.2273 0.0545  4.1701 0.0000  0.1205 0.3341
#> 
#> $`15`
#>          interval     est     se       z      p    2.5%  97.5%
#> total          15  0.0618 0.0315  1.9637 0.0496  0.0001 0.1235
#> direct         15 -0.0075 0.0081 -0.9233 0.3558 -0.0233 0.0084
#> indirect       15  0.0693 0.0305  2.2693 0.0232  0.0094 0.1291
#> 
#> $`20`
#>          interval     est     se       z      p    2.5%  97.5%
#> total          20  0.0213 0.0179  1.1885 0.2347 -0.0138 0.0564
#> direct         20 -0.0012 0.0019 -0.6086 0.5428 -0.0049 0.0026
#> indirect       20  0.0224 0.0172  1.3025 0.1928 -0.0113 0.0562
#> 
#> 
#> $mc
#> $`5`
#>          interval     est     se     R    2.5%   97.5%
#> total           5  0.3629 0.0393 20000  0.2935  0.4472
#> direct          5 -0.2177 0.0655 20000 -0.3797 -0.1263
#> indirect        5  0.5807 0.0957 20000  0.4337  0.8067
#> 
#> $`10`
#>          interval     est     se     R    2.5%   97.5%
#> total          10  0.1822 0.0431 20000  0.1151  0.2844
#> direct         10 -0.0451 0.0384 20000 -0.1553 -0.0130
#> indirect       10  0.2273 0.0634 20000  0.1469  0.3915
#> 
#> $`15`
#>          interval     est     se     R    2.5%   97.5%
#> total          15  0.0618 0.0346 20000  0.0131  0.1480
#> direct         15 -0.0075 0.0182 20000 -0.0591 -0.0003
#> indirect       15  0.0693 0.0363 20000  0.0275  0.1672
#> 
#> $`20`
#>          interval     est     se     R    2.5%  97.5%
#> total          20  0.0213 0.0219 20000 -0.0018 0.0810
#> direct         20 -0.0012 0.0095 20000 -0.0223 0.0001
#> indirect       20  0.0224 0.0224 20000  0.0034 0.0859
#> 
#>